fix(clones): clamp same-file periodic extension + integer-only knobs
Post-extension same-stream overlap now clamps at the period boundary
instead of discarding, so 3+ tandem copies keep every instance; a
class-level union-cover pass collapses the staggered sub-pair family
the clamp would otherwise mint. minTokens/minLines/gapTokens are
integer-validated in both schema and CLI (fractional values indexed
past token-array bounds).

Noldor-FD: code-clone-detector
Noldor-Path: full-new

+++ src/core/config.ts
@@ -148,9 +148,9 @@ export type GardenConfig = z.infer<typeof gardenConfigSchema>;
  * threshold as green.
  */
 export const clonesConfigSchema = z.object({
-  minTokens: z.number().positive().optional().catch(undefined),
-  minLines: z.number().positive().optional().catch(undefined),
-  gapTokens: z.number().positive().optional().catch(undefined),
+  minTokens: z.number().int().positive().optional().catch(undefined),
+  minLines: z.number().int().positive().optional().catch(undefined),
+  gapTokens: z.number().int().positive().optional().catch(undefined),
   thresholdPct: z.number().positive().optional().catch(undefined),
 });
